backend/requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
python-dotenv
networkx
numpy
scikit-learn
sentence-transformers
fpdf2
langgraph
langchain-core
langchain-groq
faiss-cpu
neo4j
pymupdf
pdfplumber
pillow
boto3
pytest
